bug: Only run os_Update_File_System_Cache when erase is successfull
Only running this function when an erase has completed successfully that may otherwise leave the device in a state where it will reject reads, writes, flushes, etc.
This causes sanitize status to stop reporting progress when issued before starting sanitize and may affect other erases as well. Because of that it is being used more sparingly and only when we are confident the drive is in a good state to accept other LBA access commands again.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>