quick: Switching to using M_DEPRECATED_REASON to make more clear the action needed in format.h
The get_Number_Of_Supported_Sector sizes was using this to help with memory allocation since the structure used a variable length array at the end that is allocated by the caller.
That is now deprecated as it is now a fixed size of 64 to be able to hold all possible 64 formats from an NVMe drive. This function is deprecated as it now just returns 1.

Signed-off-by: Tyler Erickson <tyler.erickson@seagate.com>